One Binary To Rule Them All
