+++ crates/driver/src/run.rs
@@ -21,9 +21,12 @@ use {
     tokio::sync::oneshot,
 };
 
-pub async fn main() {
+/// The driver entry-point. This function exists in order to be able to run the
+/// driver from multiple binaries.
+pub async fn start(args: impl Iterator<Item = String>) {
     boundary::exit_process_on_panic::set_panic_hook();
-    run(std::env::args(), infra::time::Now::Real, None).await
+    let args = cli::Args::parse_from(args);
+    run_with(args, infra::time::Now::Real, None).await
 }
 
 /// This function exists to enable running the driver for testing. The
@@ -37,6 +40,16 @@ pub async fn run(
     addr_sender: Option<oneshot::Sender<SocketAddr>>,
 ) {
     let args = cli::Args::parse_from(args);
+    run_with(args, now, addr_sender).await;
+}
+
+/// Run the driver. This function exists to avoid multiple monomorphizations of
+/// the `run` code, which bloats the binaries and increases compile times.
+async fn run_with(
+    args: cli::Args,
+    now: infra::time::Now,
+    addr_sender: Option<oneshot::Sender<SocketAddr>>,
+) {
     observe::init(&args.log);
     let config = config::file::load(&args.config).await;
