+++ docker/Dockerfile.binary
@@ -13,13 +13,7 @@ FROM docker.io/debian:bullseye-slim
 
 # Handle signal handlers properly
 RUN apt-get update && apt-get install -y ca-certificates tini && apt-get clean
-COPY --from=cargo-build /src/target/release/alerter /usr/local/bin/alerter
-COPY --from=cargo-build /src/target/release/autopilot /usr/local/bin/autopilot
-COPY --from=cargo-build /src/target/release/driver /usr/local/bin/driver
-COPY --from=cargo-build /src/target/release/orderbook /usr/local/bin/orderbook
-COPY --from=cargo-build /src/target/release/refunder /usr/local/bin/refunder
-COPY --from=cargo-build /src/target/release/solver /usr/local/bin/solver
-COPY --from=cargo-build /src/target/release/solvers /usr/local/bin/solvers
+COPY --from=cargo-build /src/target/release/cow /usr/local/bin/cow
 
 CMD echo "Specify binary..."
-ENTRYPOINT ["/usr/bin/tini", "--"]
+ENTRYPOINT ["/usr/bin/tini", "--", "cow"]
